refactor: decompose Layout.tsx into Header, NavLinks, BurgerMenu, Logo, AuthButtons
