+++ src/components/Layout/BurgerMenu.tsx
@@ -1,0 +1,60 @@
+import {
+  Drawer,
+  DrawerOverlay,
+  DrawerContent,
+  DrawerHeader,
+  DrawerBody,
+  IconButton,
+  Box,
+  VStack,
+} from '@chakra-ui/react';
+import { CloseIcon } from '@chakra-ui/icons';
+import NavLinks from './NavLinks';
+import AuthButtons from './AuthButtons';
+
+interface BurgerMenuProps {
+  isOpen: boolean;
+  onClose: () => void;
+  isLoggedIn: boolean;
+  showAuthButtons?: boolean;
+}
+
+const BurgerMenu = ({
+  isOpen,
+  onClose,
+  isLoggedIn,
+  showAuthButtons,
+}: BurgerMenuProps) => {
+  return (
+    <Drawer isOpen={isOpen} placement="left" onClose={onClose}>
+      <DrawerOverlay />
+      <DrawerContent>
+        <DrawerHeader display="flex" justifyContent="flex-start" p={2}>
+          <IconButton
+            aria-label="Close menu"
+            icon={<CloseIcon />}
+            variant="ghost"
+            onClick={onClose}
+          />
+        </DrawerHeader>
+        <DrawerBody>
+          <Box mt={4}>
+            <VStack spacing={4} align="stretch">
+              <NavLinks onClick={onClose} />
+            </VStack>
+            <Box pt={4} borderTop="1px solid" borderColor="gray.200">
+              <AuthButtons
+                isLoggedIn={isLoggedIn}
+                isVertical
+                onClose={onClose}
+                showAuthButtons={showAuthButtons}
+              />
+            </Box>
+          </Box>
+        </DrawerBody>
+      </DrawerContent>
+    </Drawer>
+  );
+};
+
+export default BurgerMenu;
